pyproject.toml
[project]
name = "showcase-bot"
version = "0.1.0"
description = "Discord bot for showcase-esque channels"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "discord>=2.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
